lnwallet: extend AuxLeafStore w/ additional methods
When constructing or validating a commitment transaction, we'll need to be able to obtain the aux leaves when:
  * we're making a new state
  * we're making a resolution for an existing state
  * we're handling an on chain breach

The three top level methods now handle these three cases.

The final method `ApplyHtlcView` will be used when we've made a new state, and need to derive the opaque blob that we'll store in the db for that respective commitment.